fab/requirements.txt
langgraph>=0.1.0
langchain>=0.0.300
langchain-groq>=0.0.4
chromadb>=0.3.22
sentence-transformers>=2.2.2
numpy>=1.24
pymupdf>=1.22.0
python-dotenv>=1.0.0
pandas>=2.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
tqdm>=4.65.0
rich>=13.3.0
langchain-community
langchain-chroma
pydantic
python-dotenv
unstructured
pypdf
tiktoken
numpy
pandas
gradio